codemint-backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
pytest==6.2.5
python-dotenv==0.19.0
ipfshttpclient==0.7.0
elasticsearch==7.14.0
requests==2.26.0